requirements.txt
# GTMotion.ai LinkedIn Automation Platform Dependencies

# Core dependencies
openai>=1.54.0,<2.0.0
boto3>=1.35.0,<2.0.0
botocore>=1.35.0,<2.0.0

# Encryption for GDPR compliance
cryptography>=41.0.0,<46.0.0

# HTTP requests for API calls
requests>=2.31.0,<3.0.0

# Data handling
python-dateutil>=2.8.0,<3.0.0

# Testing dependencies
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
moto>=4.2.0,<5.0.0
unittest-xml-reporting>=3.2.0,<4.0.0

# Development dependencies
black>=23.12.0,<24.0.0
flake8>=7.0.0,<8.0.0
mypy>=1.8.0,<2.0.0

# AWS Lambda deployment
awscli>=1.35.0,<2.0.0
aws-sam-cli>=1.105.0,<2.0.0

# Optional: For local development and testing
python-dotenv>=1.0.0,<2.0.0
jupyter>=1.0.0,<2.0.0

# Type hints
types-requests>=2.31.0
